requirements.txt
pandas
python-dotenv
gitpython
cryptography
almapiwrapper
lxml
requests